DSNYMETQUA-8 Render report to plaintext string
We've got some angular code for rendering it in a browser already, but
we can't easily use that in the body of an email. So we add a toString
that formats it similarly to our webpage.

Also shuffled around the abstractions around Problem to have an extra
layer for Problems that relate to a property, so we can render this a
little nicer.