DSNYMETQUA-8 Avoid some specific false positives
Alfresco comes with some 'problems' out of the box, these aren't
necessarily abnormal, so we want to prevent these edge cases from ending
up in the report, and in case of a gray zone, add an extra message (such
as with the abeecher and mjackson users). Also restructured some of the
control flow around definition == null and value == null checking, we
don't really care if a property doesn't appear in the content model if
its value is also null (this caused the cm:source is undefined false
positive previously).