ALFREDOPS-289 dont let c4a import itself
Before this commit, care4alf would import some of its own (exported)
packages, such as eu.xenit.care4alf.smtp.dumbster. This is correct,
according to the OSGi philosophy (see section 3.6.6 of the OSGi Core
spec, version 7.0.0). However, this is intended for pure API packages,
and the goal of doing this (hotswap-kind of extensibility) would never
be relevant for care4alf and it causes warning in some other tools in
our toolset, such as Dynamic Extensions. That's why we're preventing
this specific behaviour with ;-noimport:=true.