ALFREDOPS-382 Replace all escape sequences
Previously, I was just replacing the escape sequence for space (_x0020_),
because that was the most common one and I figured it might be the only
one we would reasonably encounter. I was wrong. If a noderef's name
starts with a digit, the final section of the qname path of this noderef
will have an escape digit as it's first "character", e.g.: "123.jpg"
becomes .../cm:_x0031_23.jpg

I haven't been able to find any documentation on this kind of escaping,
so I can't know for sure whether I really cover all cases here. Off the
top of my head, I'm not covering the cases of any codepoint that needs
more than 2 bytes in UTF-8, but I don't know if those are valid in qname
paths in the first place (Alfresco doesn't deal well with 4-byte emoji).

For now this should be suitable for most cases we're likely to encounter
in the wild, and things don't really break in a very bad way in the
other case.